Fixed global, added some more features
Intake global would not have worked because of the error in setting the values within it.  Also separated left and right pneumatics, and implemented the idea of an intake with a little give so it can flex upward. this logic might change a bit,as of now it just turns to "on" while intake is running, might cause unessecary pneumatic venting though.