Added in all of the safeties
I went back into the github and found every safety subvi that we had written and put them into the dashboard. The indicators are in the operation tab. Also, all of the safeties need to be plugged into the thing that they are a safety for. It has not been tested whether the safeties work in the dashboard or not, but they have been tested previously in the robot project.